CELE-78 Separate subcommands in its own files
